fix linux compile
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2589 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/GBAPipe.cpp
@@ -167,8 +167,8 @@ namespace GBAPipe
 	void StartServer() { }
 	void ConnectAsClient() { }
 	void Stop() { }
-	void Read(u32& data){}
-	void Write(u32 data){}
+	void Read(u8& data) { }
+	void Write(u8 data) { }
 	bool IsEnabled() { return false; }
 	bool IsServer() { return false; }
 }
